Make Interval a class template
Made Interval a class template templated on its underlying type. This
was needed because EFloat uses it for the error bounds but with float
as the underlying type (this was missed when implementing EFloat).
All current clients of Interval (except for EFloat) were changed to use
Intervalf - a new type alias for Interval<Float>.