Make SampledSpectrum's fromSamples methods safer
Made these methods shrink the input samples to spans of the same length
which makes the asserts redundant.
Making the SpectrumSamples a class with a constructor handling this
shrinking would be cleaner but then the spans could be given in a
wrong order so this is a better solution given this trade-off.